requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
python-dotenv==1.0.0
aiosqlite==0.19.0
qrcode[pil]==7.4.2
pillow==10.1.0
